AREA Dream Oldale. Items are faster to get. Implemented Hidden Item Quantity. Optimized BgEvent.bgUnion.hiddenItem.hiddenItemId. Renamed Hidden Item Flags and added a count.

+++ include/global.fieldmap.h
@@ -107,7 +107,7 @@ struct BgEvent
         u8 *script;
         struct {
             u16 item;
-            u16 hiddenItemId;
+            u8 hiddenItemId;
         } hiddenItem;
         u32 secretBaseId;
     } bgUnion;
